\name{ecoforge-package}
\alias{ecoforge-package}
\alias{ecoforge}
\docType{package}
\title{
\packageTitle{ecoforge}
}
\description{
\packageDescription{ecoforge}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ecoforge}
\packageIndices{ecoforge}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ecoforge}

Maintainer: \packageMaintainer{ecoforge}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
ls()
}
